+++ addons/main/XEH_PREP.hpp
@@ -1,21 +1,21 @@
 PREP(compileSanitisedFunction);
-PREP(createResupplyGUI);
-PREP(exitUnconsciousUnit);
-PREP(getRespawnTime);
 PREP(getRole);
+PREP(globalEventJIP);
 PREP(logCuratorObjects);
 PREP(objectsCounterMissionEH);
-PREP(openACEMenu);
 PREP(pathingCondition);
 PREP(pathingStatement);
+PREP(removeGlobalEventJIP);
 PREP(sanitiseFunction);
 PREP(switchUnitStart);
 PREP(switchUnitStop);
 PREP(unitParadropContextMenu);
+
+PREP(gui_createResupply);
+PREP(gui_garrisonBuilding);
 
 // 'PREP_MP' requires 'FUNC(compileSanitisedFunction)'
 PREP_MP(init);
 
 PREP_MP(changeReason);
 PREP_MP(handleBuildingDestruction);
-PREP_MP(handleJIP);