+++ addons/main/gui.hpp
@@ -38,7 +38,6 @@ class GVAR(RscDisplay) {
                     colorBackground[] = {QUOTE(GETPRVAR('GUI_BCG_RGB_R',0.13)), QUOTE(GETPRVAR('GUI_BCG_RGB_G',0.54)), QUOTE(GETPRVAR('GUI_BCG_RGB_B',0.21)), QUOTE(GETPRVAR('GUI_BCG_RGB_A',0.8))};
                     text = CSTRING(selectMagazines);
                 };
-
                 class Background: RscText {
                     idc = -1;
                     x = 0;
@@ -47,7 +46,6 @@ class GVAR(RscDisplay) {
                     h = POS_H(18.9);
                     colorBackground[] = {0, 0, 0, 0.5};
                 };
-
                 class BackgroundCategories: RscTextMulti {
                     idc = -1;
                     x = POS_X(4);
@@ -59,8 +57,7 @@ class GVAR(RscDisplay) {
                     text = CSTRING(categories);
                     tooltip = CSTRING(categoriesDesc);
                 };
-
-                class GVAR(listCategories): RscListBox {
+                class ListCategories: RscListBox {
                     idc = IDC_LIST_CATEGORIES;
                     x = POS_X(7.8);
                     y = POS_Y(1.5);
@@ -68,8 +65,7 @@ class GVAR(RscDisplay) {
                     h = POS_H(18);
                     colorBackground[] = {0, 0, 0, 0.6};
                 };
-
-                class GVAR(listMagazines): RscListBoxMulti {
+                class ListMagazines: RscListBoxMulti {
                     idc = IDC_LIST_MAGAZINES;
                     x = POS_X(20.5);
                     y = POS_Y(1.5);
@@ -77,8 +73,7 @@ class GVAR(RscDisplay) {
                     h = POS_H(18);
                     colorBackground[] = {0, 0, 0, 0.6};
                 };
-
-                class GVAR(listSelected): RscListBoxMulti {
+                class ListSelected: RscListBoxMulti {
                     idc = IDC_LIST_SELECTED;
                     x = POS_X(34.5);
                     y = POS_Y(1.5);
@@ -86,7 +81,6 @@ class GVAR(RscDisplay) {
                     h = POS_H(18);
                     colorBackground[] = {0, 0, 0, 0.6};
                 };
-
                 class ButtonOK: RscButtonMenuOK {
                     x = POS_X(42.75);
                     y = POS_Y(20.1);
@@ -99,8 +93,7 @@ class GVAR(RscDisplay) {
                     w = POS_W(5);
                     h = POS_H(1);
                 };
-
-                class GVAR(buttonMoveInto): RscButtonMenu {
+                class ButtonMoveInto: RscButtonMenu {
                     idc = IDC_BUTTON_INTO;
                     x = POS_X(33.2);
                     y = POS_Y(10);
@@ -111,8 +104,7 @@ class GVAR(RscDisplay) {
                     text = ">";
                     tooltip = CSTRING(moveIntoSelected);
                 };
-
-                class GVAR(buttonMoveOutOf): RscButtonMenu {
+                class ButtonMoveOutOf: RscButtonMenu {
                     idc = IDC_BUTTON_OUTOF;
                     x = POS_X(33.2);
                     y = POS_Y(12);
@@ -122,8 +114,7 @@ class GVAR(RscDisplay) {
                     sytle = ST_CENTER;
                     tooltip = CSTRING(moveOutOfSelected);
                 };
-
-                class GVAR(buttonClear): ctrlButtonPicture {
+                class ButtonClear: ctrlButtonPicture {
                     idc = IDC_BUTTON_CLR;
                     x = POS_X(33.2);
                     y = POS_Y(8);
@@ -133,8 +124,7 @@ class GVAR(RscDisplay) {
                     text = "\a3\3den\data\cfg3den\history\deleteitems_ca.paa";
                     tooltip = CSTRING(clearAllDesc);
                 };
-
-                class GVAR(buttonIncrement): RscButtonMenu {
+                class ButtonIncrement: RscButtonMenu {
                     idc = IDC_BUTTON_INC;
                     x = POS_X(33.2);
                     y = POS_Y(1.5);
@@ -145,8 +135,7 @@ class GVAR(RscDisplay) {
                     text = "+";
                     tooltip = CSTRING(increaseDesc);
                 };
-
-                class GVAR(buttonDecrement): RscButtonMenu {
+                class ButtonDecrement: RscButtonMenu {
                     idc = IDC_BUTTON_DEC;
                     x = POS_X(33.2);
                     y = POS_Y(3);
@@ -158,6 +147,242 @@ class GVAR(RscDisplay) {
                     tooltip = CSTRING(decreaseDesc);
                 };
             };
+        };
+    };
+};
+
+class RscPicture;
+class RscControlsGroupNoScrollbars;
+
+class ctrlXSliderH;
+class ctrlTree;
+class ctrlListbox;
+class ctrlToolbox;
+class ctrlCheckbox;
+class ctrlButtonPictureKeepAspect;
+
+class zen_common_RscLabel;
+class zen_common_RscBackground;
+class zen_common_RscEdit;
+class zen_modules_RscSidesCombo;
+
+class zen_common_RscDisplay {
+    class controls {
+        class Title;
+        class Background;
+        class Content;
+        class ButtonOK;
+        class ButtonCancel;
+    };
+};
+class zen_modules_RscDisplay: zen_common_RscDisplay {};
+
+class GVAR(rscSpawnGarrison): zen_modules_RscDisplay {
+    onLoad = QUOTE(SETUVAR(QQGVAR(display),_this select 0));
+    class controls: controls {
+        class Title: Title {
+            text = CSTRING(garrisonBuildingModuleName);
+        };
+        class Background: Background {
+            y = POS_H(1.1);
+            h = POS_H(21.8);
+        };
+        class Content: Content {
+            h = POS_H(22.5);
+
+            class controls {
+                class SideLabel: zen_common_RscLabel {
+                    text = "$STR_eval_typeside";
+                    y = POS_H(1.5);
+                };
+                class Side: zen_modules_RscSidesCombo {
+                    idc = IDC_SPAWNGARRISON_SIDE;
+                    y = POS_H(1.5);
+                };
+                class GroupSelect: RscControlsGroupNoScrollbars {
+                    idc = -1;
+                    x = 0;
+                    y = POS_H(2.6);
+                    w = POS_W(26);
+                    h = POS_H(14.2);
+
+                    class controls {
+                        class Title: zen_common_RscLabel {
+                            text = "$STR_zen_modules_GroupSelect";
+                            w = POS_W(26);
+                        };
+                        class Background: zen_common_RscBackground {
+                            x = 0;
+                            y = POS_H(1);
+                            w = POS_W(26);
+                            h = POS_H(13.2);
+                        };
+                        class TreeMode: ctrlToolbox {
+                            idc = IDC_SPAWNGARRISON_TREE_MODE;
+                            x = POS_W(0.1);
+                            y = POS_H(1.1);
+                            w = POS_W(13);
+                            h = POS_H(1);
+                            rows = 1;
+                            columns = 2;
+                            strings[] = {"$STR_zen_common_Premade", "$STR_Radio_Custom"};
+                            colorBackground[] = {0, 0, 0, 0.7};
+                        };
+                        class TreeGroups: ctrlTree {
+                            idc = IDC_SPAWNGARRISON_TREE_GROUPS;
+                            x = POS_W(0.1);
+                            y = QUOTE(2.1 * GUI_GRID_H - pixelH);
+                            w = POS_W(13);
+                            h = POS_H(12);
+                            sizeEx = QUOTE(3.96 * (1 / (getResolution select 3)) * pixelGrid * 0.5);
+                            colorBackground[] = {0, 0, 0, 0.3};
+                            colorBorder[] = {0, 0, 0, 0};
+                            disableKeyboardSearch = 1;
+                        };
+                        class TreeUnits: TreeGroups {
+                            idc = IDC_SPAWNGARRISON_TREE_UNITS;
+                        };
+                        class Label: zen_common_RscLabel {
+                            text = "$STR_zen_modules_CurrentGroup";
+                            x = POS_W(13.2);
+                            y = POS_H(1.1);
+                            w = POS_W(12.7);
+                            colorBackground[] = {0, 0, 0, 0.7};
+                        };
+                        class UnitCount: Label {
+                            idc = IDC_SPAWNGARRISON_UNIT_COUNT;
+                            style = ST_RIGHT;
+                            text = "0";
+                            w = POS_W(11.1);
+                            colorBackground[] = {0, 0, 0, 0};
+                        };
+                        class UnitList: ctrlListbox {
+                            idc = IDC_SPAWNGARRISON_UNIT_LIST;
+                            x = POS_W(13.2);
+                            y = QUOTE(2.1 * GUI_GRID_H - pixelH);
+                            w = POS_W(12.7);
+                            h = POS_H(12);
+                            colorBackground[] = {0, 0, 0, 0.3};
+                        };
+                        class UnitIcon: RscPicture {
+                            idc = -1;
+                            text = ICON_PERSON;
+                            x = POS_W(24);
+                            y = POS_H(1.1);
+                            w = POS_W(1);
+                            h = POS_H(1);
+                        };
+                        class UnitClear: ctrlButtonPictureKeepAspect {
+                            idc = IDC_SPAWNGARRISON_UNIT_CLEAR;
+                            text = "\a3\3den\data\cfg3den\history\deleteitems_ca.paa";
+                            x = POS_W(24.9);
+                            y = POS_H(1.1);
+                            w = POS_W(1);
+                            h = POS_H(1);
+                            colorBackground[] = {0, 0, 0, 0};
+                            offsetPressedX = 0;
+                            offsetPressedY = 0;
+                        };
+                    };
+                };
+                class Properties: RscControlsGroupNoScrollbars {
+                    idc = -1;
+                    x = 0;
+                    y = POS_H(16.9);
+                    w = POS_W(26);
+                    h = POS_H(5.5);
+
+                    class controls {
+                        class Title: zen_common_RscLabel {
+                            text = "$STR_A3_RscDisplayLogin_Properties";
+                            w = POS_W(26);
+                        };
+                        class Background: zen_common_RscBackground {
+                            x = 0;
+                            y = POS_H(1);
+                            w = POS_W(26);
+                            h = POS_H(6.7);
+                        };
+                        class DynamicSimulationLabel: zen_common_RscLabel {
+                            text = "$STR_3DEN_DynamicSimulation_textSingular";
+                            x = POS_W(3);
+                            y = POS_H(1.1);
+                            w = POS_W(8.9);
+                            colorBackground[] = {0, 0, 0, 0.7};
+                        };
+                        class DynamicSimulation: ctrlCheckbox {
+                            idc = IDC_SPAWNGARRISON_DYNAMIC_SIMULATION;
+                            x = POS_W(12);
+                            y = POS_H(1.1);
+                            w = POS_W(1);
+                            h = POS_H(1);
+                        };
+                        class TriggerLabel: zen_common_RscLabel {
+                            text = "$STR_3DEN_CfgVehicles_EmptyDetector";
+                            x = POS_W(3);
+                            y = POS_H(2.2);
+                            w = POS_W(8.9);
+                            colorBackground[] = {0, 0, 0, 0.7};
+                            tooltip = "$STR_3den_attributes_triggeractivation_anyplayer_text";
+                        };
+                        class Trigger: ctrlCheckbox {
+                            idc = IDC_SPAWNGARRISON_TRIGGER;
+                            x = POS_W(12);
+                            y = POS_H(2.2);
+                            w = POS_W(1);
+                            h = POS_H(1);
+                        };
+                        class TriggerRadiusLabel: zen_common_RscLabel {
+                            text = "$STR_3den_trigger_attribute_size_displayname";
+                            x = POS_W(3);
+                            y = POS_H(3.3);
+                            w = POS_W(8.9);
+                            colorBackground[] = {0, 0, 0, 0.7};
+                        };
+                        class TriggerRadius: ctrlXSliderH {
+                            idc = IDC_SPAWNGARRISON_TRIGGER_RADIUS;
+                            x = POS_W(12);
+                            y = POS_H(3.3);
+                            w = POS_W(11);
+                            h = POS_H(1);
+                            lineSize = 25;
+                            pageSize = 5;
+                            sliderStep = 5;
+                            sliderRange[] = {0, 500};
+                            sliderPosition = 0;
+                        };
+                        class UnitBehaviourLabel: zen_common_RscLabel {
+                            text = "$STR_zen_modules_UnitBehaviour";
+                            x = POS_W(3);
+                            y = POS_H(4.4);
+                            w = POS_W(8.9);
+                            colorBackground[] = {0, 0, 0, 0.7};
+                        };
+                        class UnitBehaviour: ctrlToolbox {
+                            idc = IDC_SPAWNGARRISON_UNIT_BEHAVIOUR;
+                            x = POS_W(12);
+                            y = POS_H(4.4);
+                            w = POS_W(11);
+                            h = POS_H(1);
+                            rows = 1;
+                            columns = 4;
+                            strings[] = {
+                                "$STR_Disp_Default",
+                                "$STR_zen_common_Relaxed",
+                                "$STR_zen_common_Cautious",
+                                "$STR_Combat"
+                            };
+                            colorBackground[] = {0, 0, 0, 0.7};
+                        };
+                    };
+                };
+            };
+        };
+        class ButtonOK: ButtonOK {
+            y = POS_H(23);
+        };
+        class ButtonCancel: ButtonCancel {
+            y = POS_H(23);
         };
     };
 };